alterações do feedback recebido 23/05, básicas de categoria e insumos e fundamental medicar/vacinar

+++ src/main/java/sosanimais/com/example/app/model/entity/Categoria.java
@@ -1,21 +1,21 @@
 package sosanimais.com.example.app.model.entity;
 
 public class Categoria {
-    private int id;
+    private Long id;
     private String nome;
     private String descricao;
 
-    public Categoria(int id, String nome, String descricao) {
+    public Categoria(Long id, String nome, String descricao) {
         this.id = id;
         this.nome = nome;
         this.descricao = descricao;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
